+++ CHANGES.md
@@ -25,6 +25,7 @@ Changes in 4.8
 - Add TranscriptPDFHeader() & TranscriptPDFFooter() functions in Transcripts.fnc.php
 - Add Transcript PDF Header & Footer action hooks in Transcripts.fnc.php
 - Search Parents by Student Grade Level in Search.fnc.php & GetStaffList.fnc.php, sponsored by Santa Cecilia school
+- Add link to associated Student / Parent Info in AddStudents.php & AddUsers.php
 
 Changes in 4.7.1
 ----------------